Refatoraçao iniciada, endpoints do usuario reorganizados.
