+++ semana18/cookenu/src/data/recipeQueries.ts
@@ -1,0 +1,1 @@
+import connection from "./connection";
