+++ semana18/cookenu/src/types/authentication.ts
@@ -1,0 +1,4 @@
+export type authenticationData = {
+  id: string;
+  role: string;
+};
