+++ semana18/cookenu/src/types/feed.ts
@@ -1,0 +1,8 @@
+export type feedData = {
+  id: string;
+  title: string;
+  description: string;
+  user_id: string;
+  name: string;
+  created_at: Date;
+};
