+++ semana18/cookenu/src/types/follow.ts
@@ -1,0 +1,4 @@
+export type follow = {
+  followed_id: string;
+  follower_id: string;
+};
