+++ semana18/cookenu/src/types/recipe.ts
@@ -1,0 +1,6 @@
+export type recipe = {
+  id: string;
+  title: string;
+  description: string;
+  user_id: string;
+};
