+++ semana18/cookenu/src/types/user.ts
@@ -1,0 +1,12 @@
+export enum USER_ROLES {
+  ADMIN = "ADMIN",
+  NORMAL = "NORMAL",
+}
+
+export type user = {
+  id: string;
+  name: string;
+  email: string;
+  password: string;
+  role: string;
+};
